Commit1
Commit including topic frequency counts per year of available ECAI proceedings and nationalities of authors. The predictive model is included which frames the problem as a time-series forecasting problem - based on the previous years topic frequency, we predict the following year topic frequency of ECAI 2025. Additionally, an evaluation of the nationalities of the authors of ECAI 2024 proceedings are included.